:art: Refactor LeftSidebarContent out of Dashboard
