+++ src/pages/dashboard/DashboardLayout.js
@@ -3,26 +3,12 @@
 // Libraries
 import React, { useState, useEffect } from 'react';
 import Sidebar from 'react-sidebar';
-import RoundedImage from 'react-rounded-image';
-// import { connect } from 'react-redux';
 
-// import { fetchProfile } from '../../modules/dashboard/dashboardActions';
-// assets
-import icons from '../../assets/icons';
-import profileDefault from '../../assets/user_profile_default.jpg';
-
-// Styled
-import * as c from '../../styles/variables/colours';
 import * as g from '../../styles/variables/global';
-import { P, H1 } from '../../styles/typography';
+import LeftSideBar from '../../components/leftSideBar/LeftSideBar';
 import {
-  GrowSpace,
   HamburgerButton,
-  MenuBox,
   DashboardContainer,
-  ProfileImageDiv,
-  Item,
-  SidebarBody,
   sideBarStyle,
   sideBarRootStyle,
   MainContent,
@@ -50,7 +36,7 @@ export default function DashboardLayout(props) {
   return (
     <DashboardContainer>
       <Sidebar
-        sidebar={<SideBarContent name={user.credentials.full_name} />}
+        sidebar={<LeftSideBar name={user.credentials.full_name} />}
         open={sidebarOpen}
         onSetOpen={setSidebarOpen}
         docked={sideBarDocked}
@@ -73,45 +59,3 @@ export default function DashboardLayout(props) {
     </DashboardContainer>
   );
 }
-
-const SideBarContent = ({ name }) => {
-  return (
-    <SidebarBody>
-      <ProfileImageDiv>
-        <RoundedImage
-          // eslint-disable-next-line max-len
-          image={profileDefault}
-          alt="User's profile"
-          imageHeight="150"
-          imageWidth="150"
-          roundedSize="1"
-          roundedColor="#FFF"
-        />
-
-        <H1 REGULAR color={c.WHITE}>
-          {name}
-        </H1>
-      </ProfileImageDiv>
-
-      <MenuBox>
-        <MenuItem img={icons.DashboardIcon} text="Dashboard" />
-        <MenuItem img={icons.ProfileIcon} text="Profile" />
-        <MenuItem img={icons.AddDecksIcon} text="Add Decks" />
-        <MenuItem img={icons.LibraryIcon} text="Deck Library" />
-        <GrowSpace />
-        <MenuItem img={icons.SettingsIcon} text="Settings" />
-      </MenuBox>
-    </SidebarBody>
-  );
-};
-
-const MenuItem = ({ img, text, route = '/' }) => {
-  return (
-    <Item to={route}>
-      <img src={img} alt="" />
-      <P BRAND color={c.WHITE}>
-        {text}
-      </P>
-    </Item>
-  );
-};
